Add full PS1 config to ps1_notifications.sh
This variation of PS1 is with the bracket, so it looks like this:

┌─user@host~/scripts/config (master *+)
└─$

README notification goes after the git information. Change the color of
the Z symbol to be less distracting.  Sad smiley goes before the bottom
bracket, moving the dollar to the right:

┌─user@host~/scripts/config (master *+)
☹ └─$

which is a good visual cue.